backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.14.1
python-dotenv==1.1.1
python-nmap==0.7.1
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
